Add `maze.png`
This is a photo of the marble maze taken with the webcam we plan to use, in approximately the right
position. This allows for development of the computer vision side of the planning code.

Adding binary files to git like this is usually questionable practice. I'm justifying this as being
acceptable due to it being a single reasonably small image and vital for running the code that will
follow - keeping the image elsewhere would be worse.